Made progress on Prices view model
