+++ src/Sidekick.UI/ApplicationLogs/IApplicationLogViewModel.cs
@@ -1,9 +1,10 @@
 using System.Collections.ObjectModel;
+using System.ComponentModel;
 using Sidekick.Core.Loggers;
 
 namespace Sidekick.UI.ApplicationLogs
 {
-    public interface IApplicationLogViewModel
+    public interface IApplicationLogViewModel : INotifyPropertyChanged
     {
         ObservableCollection<Log> Logs { get; }
     }
