+++ src/Sidekick.UI/Settings/SettingsViewModel.cs
@@ -1,6 +1,6 @@
 using System;
+using System.ComponentModel;
 using System.Linq;
-using PropertyChanged;
 using Sidekick.Business.Leagues;
 using Sidekick.Core.Natives;
 using Sidekick.Core.Settings;
@@ -9,13 +9,14 @@ using Sidekick.UI.Helpers;
 
 namespace Sidekick.UI.Settings
 {
-    [AddINotifyPropertyChangedInterface]
     public class SettingsViewModel : ISettingsViewModel, IDisposable
     {
         private readonly IUILanguageProvider uiLanguageProvider;
         private readonly SidekickSettings sidekickSettings;
         private readonly INativeKeyboard nativeKeyboard;
         private readonly IKeybindEvents keybindEvents;
+
+        public event PropertyChangedEventHandler PropertyChanged;
 
         public SettingsViewModel(IUILanguageProvider uiLanguageProvider,
             SidekickSettings sidekickSettings,
